deployment and admin UI setup
